renderPoll function written. trying to figure out how renderOption function fits in
