+++ index.html
@@ -36,12 +36,7 @@
 
           <!-- CURRENT POLL SECTION -->
           <section class="current-poll-section" id="current-poll-section">
-            <!-- <h2 id="current-poll-question">###</h2> -->
-            <!-- <span class="current-option" id="current-optionA">###</span> -->
-            <!-- <span class="current-option" id="current-optionB">###</span> -->
-
-            <!-- <span class="current-result" id="resultA">###</span> -->
-            <!-- <span class="current-result" id="resultB">###</span> -->
+            <!-- Current Poll item elements will be created in DOM -->
           </section>
 
           <section class="button-section">
@@ -58,7 +53,7 @@
           <section class="past-results-section">
             <h2>Past Results</h2>
             <div id="closed-poll-container">
-              <!-- Closed Poll item elements will be created in JSON -->
+              <!-- Closed Poll item elements will be created in DOM -->
             </div>
           </section>
     </main>
